+++ lib/src/config.dart
@@ -1,5 +1,5 @@
 class LazyLocConfig {
   static const String targetDir = 'assets/translations';
-  static const List<String> defaultTargetLangs = ['ko', 'en'];
+  static const List<String> defaultTargetLangs = ['en'];
   static const String backupDirName = '_backup';
 }