+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.1+1
+
+* Fix: feat: Replace print statements with Logger utility
+* Chore: Update default target language from ['ko', 'en'] to ['en'] only.
+
 ## 0.0.1
 
 * Initial release